Run `build` after polyfill-service install
The output when we try to boot the version without `npm run build` tells
us we need to do this.

In order to get `npm run build` to work, we needed a new version of NPM
and to install all the dependencies.

It would be cool if we could use `npm ci`, but we need the
`devDependencies` to do `npm run build`, so we have to use `npm
install`. Also, because we've set `NODE_ENV` to `production`, `npm
install` won't install the dev dependencies by default, so we have to
use `--production=false`.